requirements.txt
# SkyLore - 依存パッケージ

# Web UI
streamlit>=1.28.0

# OpenAI API
openai>=1.0.0

# ユーティリティ
python-dotenv>=1.0.0

# 形態素解析（オプション：友達がインデックス改良時に使用）
# mecab-python3>=1.0.0
# unidic-lite>=1.0.0
